Former-commit-id: d917f226cd5ddca9fe2d2650eae4fb428eefb080

+++ PLATFORM/Modules/DynamicExpression/VirtoCommerce.DynamicExpressionModule.Data/Promotion/Conditions/CartConditions/ConditionAtNumItemsOfCategoryAreInCart.cs
@@ -9,7 +9,7 @@ namespace VirtoCommerce.DynamicExpressionModule.Data.Promotion
 	public class ConditionAtNumItemsInCategoryAreInCart : ConditionBase, IConditionExpression
 	{
 	
-		public decimal NumItem { get; set; }
+		public int NumItem { get; set; }
 	
 		public bool Exactly { get; set; }
